+++ project/vivado_2015.4/ipcores/wrap_switch_core.vhd
@@ -56,6 +56,7 @@ entity wrap_switch_core is
     p00_rx_write    : in  std_logic;
     p00_rx_error    : in  std_logic;
     p00_rx_rate     : in  std_logic_vector(15 downto 0);
+    p00_rx_status   : in  std_logic_vector(7 downto 0);
     p00_rx_reset    : in  std_logic;
     p00_tx_clk      : in  std_logic;
     p00_tx_data     : out std_logic_vector(7 downto 0);
@@ -71,6 +72,7 @@ entity wrap_switch_core is
     p01_rx_write    : in  std_logic;
     p01_rx_error    : in  std_logic;
     p01_rx_rate     : in  std_logic_vector(15 downto 0);
+    p01_rx_status   : in  std_logic_vector(7 downto 0);
     p01_rx_reset    : in  std_logic;
     p01_tx_clk      : in  std_logic;
     p01_tx_data     : out std_logic_vector(7 downto 0);
@@ -86,6 +88,7 @@ entity wrap_switch_core is
     p02_rx_write    : in  std_logic;
     p02_rx_error    : in  std_logic;
     p02_rx_rate     : in  std_logic_vector(15 downto 0);
+    p02_rx_status   : in  std_logic_vector(7 downto 0);
     p02_rx_reset    : in  std_logic;
     p02_tx_clk      : in  std_logic;
     p02_tx_data     : out std_logic_vector(7 downto 0);
@@ -101,6 +104,7 @@ entity wrap_switch_core is
     p03_rx_write    : in  std_logic;
     p03_rx_error    : in  std_logic;
     p03_rx_rate     : in  std_logic_vector(15 downto 0);
+    p03_rx_status   : in  std_logic_vector(7 downto 0);
     p03_rx_reset    : in  std_logic;
     p03_tx_clk      : in  std_logic;
     p03_tx_data     : out std_logic_vector(7 downto 0);
@@ -116,6 +120,7 @@ entity wrap_switch_core is
     p04_rx_write    : in  std_logic;
     p04_rx_error    : in  std_logic;
     p04_rx_rate     : in  std_logic_vector(15 downto 0);
+    p04_rx_status   : in  std_logic_vector(7 downto 0);
     p04_rx_reset    : in  std_logic;
     p04_tx_clk      : in  std_logic;
     p04_tx_data     : out std_logic_vector(7 downto 0);
@@ -131,6 +136,7 @@ entity wrap_switch_core is
     p05_rx_write    : in  std_logic;
     p05_rx_error    : in  std_logic;
     p05_rx_rate     : in  std_logic_vector(15 downto 0);
+    p05_rx_status   : in  std_logic_vector(7 downto 0);
     p05_rx_reset    : in  std_logic;
     p05_tx_clk      : in  std_logic;
     p05_tx_data     : out std_logic_vector(7 downto 0);
@@ -146,6 +152,7 @@ entity wrap_switch_core is
     p06_rx_write    : in  std_logic;
     p06_rx_error    : in  std_logic;
     p06_rx_rate     : in  std_logic_vector(15 downto 0);
+    p06_rx_status   : in  std_logic_vector(7 downto 0);
     p06_rx_reset    : in  std_logic;
     p06_tx_clk      : in  std_logic;
     p06_tx_data     : out std_logic_vector(7 downto 0);
@@ -161,6 +168,7 @@ entity wrap_switch_core is
     p07_rx_write    : in  std_logic;
     p07_rx_error    : in  std_logic;
     p07_rx_rate     : in  std_logic_vector(15 downto 0);
+    p07_rx_status   : in  std_logic_vector(7 downto 0);
     p07_rx_reset    : in  std_logic;
     p07_tx_clk      : in  std_logic;
     p07_tx_data     : out std_logic_vector(7 downto 0);
@@ -176,6 +184,7 @@ entity wrap_switch_core is
     p08_rx_write    : in  std_logic;
     p08_rx_error    : in  std_logic;
     p08_rx_rate     : in  std_logic_vector(15 downto 0);
+    p08_rx_status   : in  std_logic_vector(7 downto 0);
     p08_rx_reset    : in  std_logic;
     p08_tx_clk      : in  std_logic;
     p08_tx_data     : out std_logic_vector(7 downto 0);
@@ -191,6 +200,7 @@ entity wrap_switch_core is
     p09_rx_write    : in  std_logic;
     p09_rx_error    : in  std_logic;
     p09_rx_rate     : in  std_logic_vector(15 downto 0);
+    p09_rx_status   : in  std_logic_vector(7 downto 0);
     p09_rx_reset    : in  std_logic;
     p09_tx_clk      : in  std_logic;
     p09_tx_data     : out std_logic_vector(7 downto 0);
@@ -206,6 +216,7 @@ entity wrap_switch_core is
     p10_rx_write    : in  std_logic;
     p10_rx_error    : in  std_logic;
     p10_rx_rate     : in  std_logic_vector(15 downto 0);
+    p10_rx_status   : in  std_logic_vector(7 downto 0);
     p10_rx_reset    : in  std_logic;
     p10_tx_clk      : in  std_logic;
     p10_tx_data     : out std_logic_vector(7 downto 0);
@@ -221,6 +232,7 @@ entity wrap_switch_core is
     p11_rx_write    : in  std_logic;
     p11_rx_error    : in  std_logic;
     p11_rx_rate     : in  std_logic_vector(15 downto 0);
+    p11_rx_status   : in  std_logic_vector(7 downto 0);
     p11_rx_reset    : in  std_logic;
     p11_tx_clk      : in  std_logic;
     p11_tx_data     : out std_logic_vector(7 downto 0);
@@ -236,6 +248,7 @@ entity wrap_switch_core is
     p12_rx_write    : in  std_logic;
     p12_rx_error    : in  std_logic;
     p12_rx_rate     : in  std_logic_vector(15 downto 0);
+    p12_rx_status   : in  std_logic_vector(7 downto 0);
     p12_rx_reset    : in  std_logic;
     p12_tx_clk      : in  std_logic;
     p12_tx_data     : out std_logic_vector(7 downto 0);
@@ -251,6 +264,7 @@ entity wrap_switch_core is
     p13_rx_write    : in  std_logic;
     p13_rx_error    : in  std_logic;
     p13_rx_rate     : in  std_logic_vector(15 downto 0);
+    p13_rx_status   : in  std_logic_vector(7 downto 0);
     p13_rx_reset    : in  std_logic;
     p13_tx_clk      : in  std_logic;
     p13_tx_data     : out std_logic_vector(7 downto 0);
@@ -266,6 +280,7 @@ entity wrap_switch_core is
     p14_rx_write    : in  std_logic;
     p14_rx_error    : in  std_logic;
     p14_rx_rate     : in  std_logic_vector(15 downto 0);
+    p14_rx_status   : in  std_logic_vector(7 downto 0);
     p14_rx_reset    : in  std_logic;
     p14_tx_clk      : in  std_logic;
     p14_tx_data     : out std_logic_vector(7 downto 0);
@@ -281,6 +296,7 @@ entity wrap_switch_core is
     p15_rx_write    : in  std_logic;
     p15_rx_error    : in  std_logic;
     p15_rx_rate     : in  std_logic_vector(15 downto 0);
+    p15_rx_status   : in  std_logic_vector(7 downto 0);
     p15_rx_reset    : in  std_logic;
     p15_tx_clk      : in  std_logic;
     p15_tx_data     : out std_logic_vector(7 downto 0);
@@ -296,6 +312,7 @@ entity wrap_switch_core is
     p16_rx_write    : in  std_logic;
     p16_rx_error    : in  std_logic;
     p16_rx_rate     : in  std_logic_vector(15 downto 0);
+    p16_rx_status   : in  std_logic_vector(7 downto 0);
     p16_rx_reset    : in  std_logic;
     p16_tx_clk      : in  std_logic;
     p16_tx_data     : out std_logic_vector(7 downto 0);
@@ -311,6 +328,7 @@ entity wrap_switch_core is
     p17_rx_write    : in  std_logic;
     p17_rx_error    : in  std_logic;
     p17_rx_rate     : in  std_logic_vector(15 downto 0);
+    p17_rx_status   : in  std_logic_vector(7 downto 0);
     p17_rx_reset    : in  std_logic;
     p17_tx_clk      : in  std_logic;
     p17_tx_data     : out std_logic_vector(7 downto 0);
@@ -326,6 +344,7 @@ entity wrap_switch_core is
     p18_rx_write    : in  std_logic;
     p18_rx_error    : in  std_logic;
     p18_rx_rate     : in  std_logic_vector(15 downto 0);
+    p18_rx_status   : in  std_logic_vector(7 downto 0);
     p18_rx_reset    : in  std_logic;
     p18_tx_clk      : in  std_logic;
     p18_tx_data     : out std_logic_vector(7 downto 0);
@@ -341,6 +360,7 @@ entity wrap_switch_core is
     p19_rx_write    : in  std_logic;
     p19_rx_error    : in  std_logic;
     p19_rx_rate     : in  std_logic_vector(15 downto 0);
+    p19_rx_status   : in  std_logic_vector(7 downto 0);
     p19_rx_reset    : in  std_logic;
     p19_tx_clk      : in  std_logic;
     p19_tx_data     : out std_logic_vector(7 downto 0);
@@ -356,6 +376,7 @@ entity wrap_switch_core is
     p20_rx_write    : in  std_logic;
     p20_rx_error    : in  std_logic;
     p20_rx_rate     : in  std_logic_vector(15 downto 0);
+    p20_rx_status   : in  std_logic_vector(7 downto 0);
     p20_rx_reset    : in  std_logic;
     p20_tx_clk      : in  std_logic;
     p20_tx_data     : out std_logic_vector(7 downto 0);
@@ -371,6 +392,7 @@ entity wrap_switch_core is
     p21_rx_write    : in  std_logic;
     p21_rx_error    : in  std_logic;
     p21_rx_rate     : in  std_logic_vector(15 downto 0);
+    p21_rx_status   : in  std_logic_vector(7 downto 0);
     p21_rx_reset    : in  std_logic;
     p21_tx_clk      : in  std_logic;
     p21_tx_data     : out std_logic_vector(7 downto 0);
@@ -386,6 +408,7 @@ entity wrap_switch_core is
     p22_rx_write    : in  std_logic;
     p22_rx_error    : in  std_logic;
     p22_rx_rate     : in  std_logic_vector(15 downto 0);
+    p22_rx_status   : in  std_logic_vector(7 downto 0);
     p22_rx_reset    : in  std_logic;
     p22_tx_clk      : in  std_logic;
     p22_tx_data     : out std_logic_vector(7 downto 0);
@@ -401,6 +424,7 @@ entity wrap_switch_core is
     p23_rx_write    : in  std_logic;
     p23_rx_error    : in  std_logic;
     p23_rx_rate     : in  std_logic_vector(15 downto 0);
+    p23_rx_status   : in  std_logic_vector(7 downto 0);
     p23_rx_reset    : in  std_logic;
     p23_tx_clk      : in  std_logic;
     p23_tx_data     : out std_logic_vector(7 downto 0);
@@ -416,6 +440,7 @@ entity wrap_switch_core is
     p24_rx_write    : in  std_logic;
     p24_rx_error    : in  std_logic;
     p24_rx_rate     : in  std_logic_vector(15 downto 0);
+    p24_rx_status   : in  std_logic_vector(7 downto 0);
     p24_rx_reset    : in  std_logic;
     p24_tx_clk      : in  std_logic;
     p24_tx_data     : out std_logic_vector(7 downto 0);
@@ -431,6 +456,7 @@ entity wrap_switch_core is
     p25_rx_write    : in  std_logic;
     p25_rx_error    : in  std_logic;
     p25_rx_rate     : in  std_logic_vector(15 downto 0);
+    p25_rx_status   : in  std_logic_vector(7 downto 0);
     p25_rx_reset    : in  std_logic;
     p25_tx_clk      : in  std_logic;
     p25_tx_data     : out std_logic_vector(7 downto 0);
@@ -446,6 +472,7 @@ entity wrap_switch_core is
     p26_rx_write    : in  std_logic;
     p26_rx_error    : in  std_logic;
     p26_rx_rate     : in  std_logic_vector(15 downto 0);
+    p26_rx_status   : in  std_logic_vector(7 downto 0);
     p26_rx_reset    : in  std_logic;
     p26_tx_clk      : in  std_logic;
     p26_tx_data     : out std_logic_vector(7 downto 0);
@@ -461,6 +488,7 @@ entity wrap_switch_core is
     p27_rx_write    : in  std_logic;
     p27_rx_error    : in  std_logic;
     p27_rx_rate     : in  std_logic_vector(15 downto 0);
+    p27_rx_status   : in  std_logic_vector(7 downto 0);
     p27_rx_reset    : in  std_logic;
     p27_tx_clk      : in  std_logic;
     p27_tx_data     : out std_logic_vector(7 downto 0);
@@ -476,6 +504,7 @@ entity wrap_switch_core is
     p28_rx_write    : in  std_logic;
     p28_rx_error    : in  std_logic;
     p28_rx_rate     : in  std_logic_vector(15 downto 0);
+    p28_rx_status   : in  std_logic_vector(7 downto 0);
     p28_rx_reset    : in  std_logic;
     p28_tx_clk      : in  std_logic;
     p28_tx_data     : out std_logic_vector(7 downto 0);
@@ -491,6 +520,7 @@ entity wrap_switch_core is
     p29_rx_write    : in  std_logic;
     p29_rx_error    : in  std_logic;
     p29_rx_rate     : in  std_logic_vector(15 downto 0);
+    p29_rx_status   : in  std_logic_vector(7 downto 0);
     p29_rx_reset    : in  std_logic;
     p29_tx_clk      : in  std_logic;
     p29_tx_data     : out std_logic_vector(7 downto 0);
@@ -506,6 +536,7 @@ entity wrap_switch_core is
     p30_rx_write    : in  std_logic;
     p30_rx_error    : in  std_logic;
     p30_rx_rate     : in  std_logic_vector(15 downto 0);
+    p30_rx_status   : in  std_logic_vector(7 downto 0);
     p30_rx_reset    : in  std_logic;
     p30_tx_clk      : in  std_logic;
     p30_tx_data     : out std_logic_vector(7 downto 0);
@@ -521,6 +552,7 @@ entity wrap_switch_core is
     p31_rx_write    : in  std_logic;
     p31_rx_error    : in  std_logic;
     p31_rx_rate     : in  std_logic_vector(15 downto 0);
+    p31_rx_status   : in  std_logic_vector(7 downto 0);
     p31_rx_reset    : in  std_logic;
     p31_tx_clk      : in  std_logic;
     p31_tx_data     : out std_logic_vector(7 downto 0);
@@ -580,6 +612,7 @@ gen_p00 : if (PORT_COUNT > 0) generate
     rx_data(0).write    <= p00_rx_write;
     rx_data(0).rxerr    <= p00_rx_error;
     rx_data(0).rate     <= p00_rx_rate;
+    rx_data(0).status   <= p00_rx_status;
     rx_data(0).reset_p  <= p00_rx_reset;
     tx_ctrl(0).clk      <= p00_tx_clk;
     tx_ctrl(0).ready    <= p00_tx_ready;
@@ -603,6 +636,7 @@ gen_p01 : if (PORT_COUNT > 1) generate
     rx_data(1).write    <= p01_rx_write;
     rx_data(1).rxerr    <= p01_rx_error;
     rx_data(1).rate     <= p01_rx_rate;
+    rx_data(1).status   <= p01_rx_status;
     rx_data(1).reset_p  <= p01_rx_reset;
     tx_ctrl(1).clk      <= p01_tx_clk;
     tx_ctrl(1).ready    <= p01_tx_ready;
@@ -626,6 +660,7 @@ gen_p02 : if (PORT_COUNT > 2) generate
     rx_data(2).write    <= p02_rx_write;
     rx_data(2).rxerr    <= p02_rx_error;
     rx_data(2).rate     <= p02_rx_rate;
+    rx_data(2).status   <= p02_rx_status;
     rx_data(2).reset_p  <= p02_rx_reset;
     tx_ctrl(2).clk      <= p02_tx_clk;
     tx_ctrl(2).ready    <= p02_tx_ready;
@@ -649,6 +684,7 @@ gen_p03 : if (PORT_COUNT > 3) generate
     rx_data(3).write    <= p03_rx_write;
     rx_data(3).rxerr    <= p03_rx_error;
     rx_data(3).rate     <= p03_rx_rate;
+    rx_data(3).status   <= p03_rx_status;
     rx_data(3).reset_p  <= p03_rx_reset;
     tx_ctrl(3).clk      <= p03_tx_clk;
     tx_ctrl(3).ready    <= p03_tx_ready;
@@ -672,6 +708,7 @@ gen_p04 : if (PORT_COUNT > 4) generate
     rx_data(4).write    <= p04_rx_write;
     rx_data(4).rxerr    <= p04_rx_error;
     rx_data(4).rate     <= p04_rx_rate;
+    rx_data(4).status   <= p04_rx_status;
     rx_data(4).reset_p  <= p04_rx_reset;
     tx_ctrl(4).clk      <= p04_tx_clk;
     tx_ctrl(4).ready    <= p04_tx_ready;
@@ -695,6 +732,7 @@ gen_p05 : if (PORT_COUNT > 5) generate
     rx_data(5).write    <= p05_rx_write;
     rx_data(5).rxerr    <= p05_rx_error;
     rx_data(5).rate     <= p05_rx_rate;
+    rx_data(5).status   <= p05_rx_status;
     rx_data(5).reset_p  <= p05_rx_reset;
     tx_ctrl(5).clk      <= p05_tx_clk;
     tx_ctrl(5).ready    <= p05_tx_ready;
@@ -718,6 +756,7 @@ gen_p06 : if (PORT_COUNT > 6) generate
     rx_data(6).write    <= p06_rx_write;
     rx_data(6).rxerr    <= p06_rx_error;
     rx_data(6).rate     <= p06_rx_rate;
+    rx_data(6).status   <= p06_rx_status;
     rx_data(6).reset_p  <= p06_rx_reset;
     tx_ctrl(6).clk      <= p06_tx_clk;
     tx_ctrl(6).ready    <= p06_tx_ready;
@@ -741,6 +780,7 @@ gen_p07 : if (PORT_COUNT > 7) generate
     rx_data(7).write    <= p07_rx_write;
     rx_data(7).rxerr    <= p07_rx_error;
     rx_data(7).rate     <= p07_rx_rate;
+    rx_data(7).status   <= p07_rx_status;
     rx_data(7).reset_p  <= p07_rx_reset;
     tx_ctrl(7).clk      <= p07_tx_clk;
     tx_ctrl(7).ready    <= p07_tx_ready;
@@ -764,6 +804,7 @@ gen_p08 : if (PORT_COUNT > 8) generate
     rx_data(8).write    <= p08_rx_write;
     rx_data(8).rxerr    <= p08_rx_error;
     rx_data(8).rate     <= p08_rx_rate;
+    rx_data(8).status   <= p08_rx_status;
     rx_data(8).reset_p  <= p08_rx_reset;
     tx_ctrl(8).clk      <= p08_tx_clk;
     tx_ctrl(8).ready    <= p08_tx_ready;
@@ -787,6 +828,7 @@ gen_p09 : if (PORT_COUNT > 9) generate
     rx_data(9).write    <= p09_rx_write;
     rx_data(9).rxerr    <= p09_rx_error;
     rx_data(9).rate     <= p09_rx_rate;
+    rx_data(9).status   <= p09_rx_status;
     rx_data(9).reset_p  <= p09_rx_reset;
     tx_ctrl(9).clk      <= p09_tx_clk;
     tx_ctrl(9).ready    <= p09_tx_ready;
@@ -810,6 +852,7 @@ gen_p10 : if (PORT_COUNT > 10) generate
     rx_data(10).write   <= p10_rx_write;
     rx_data(10).rxerr   <= p10_rx_error;
     rx_data(10).rate    <= p10_rx_rate;
+    rx_data(10).status  <= p10_rx_status;
     rx_data(10).reset_p <= p10_rx_reset;
     tx_ctrl(10).clk     <= p10_tx_clk;
     tx_ctrl(10).ready   <= p10_tx_ready;
@@ -833,6 +876,7 @@ gen_p11 : if (PORT_COUNT > 11) generate
     rx_data(11).write   <= p11_rx_write;
     rx_data(11).rxerr   <= p11_rx_error;
     rx_data(11).rate    <= p11_rx_rate;
+    rx_data(11).status  <= p11_rx_status;
     rx_data(11).reset_p <= p11_rx_reset;
     tx_ctrl(11).clk     <= p11_tx_clk;
     tx_ctrl(11).ready   <= p11_tx_ready;
@@ -856,6 +900,7 @@ gen_p12 : if (PORT_COUNT > 12) generate
     rx_data(12).write   <= p12_rx_write;
     rx_data(12).rxerr   <= p12_rx_error;
     rx_data(12).rate    <= p12_rx_rate;
+    rx_data(12).status  <= p12_rx_status;
     rx_data(12).reset_p <= p12_rx_reset;
     tx_ctrl(12).clk     <= p12_tx_clk;
     tx_ctrl(12).ready   <= p12_tx_ready;
@@ -879,6 +924,7 @@ gen_p13 : if (PORT_COUNT > 13) generate
     rx_data(13).write   <= p13_rx_write;
     rx_data(13).rxerr   <= p13_rx_error;
     rx_data(13).rate    <= p13_rx_rate;
+    rx_data(13).status  <= p13_rx_status;
     rx_data(13).reset_p <= p13_rx_reset;
     tx_ctrl(13).clk     <= p13_tx_clk;
     tx_ctrl(13).ready   <= p13_tx_ready;
@@ -902,6 +948,7 @@ gen_p14 : if (PORT_COUNT > 14) generate
     rx_data(14).write   <= p14_rx_write;
     rx_data(14).rxerr   <= p14_rx_error;
     rx_data(14).rate    <= p14_rx_rate;
+    rx_data(14).status  <= p14_rx_status;
     rx_data(14).reset_p <= p14_rx_reset;
     tx_ctrl(14).clk     <= p14_tx_clk;
     tx_ctrl(14).ready   <= p14_tx_ready;
@@ -925,6 +972,7 @@ gen_p15 : if (PORT_COUNT > 15) generate
     rx_data(15).write   <= p15_rx_write;
     rx_data(15).rxerr   <= p15_rx_error;
     rx_data(15).rate    <= p15_rx_rate;
+    rx_data(15).status  <= p15_rx_status;
     rx_data(15).reset_p <= p15_rx_reset;
     tx_ctrl(15).clk     <= p15_tx_clk;
     tx_ctrl(15).ready   <= p15_tx_ready;
@@ -948,6 +996,7 @@ gen_p16 : if (PORT_COUNT > 16) generate
     rx_data(16).write   <= p16_rx_write;
     rx_data(16).rxerr   <= p16_rx_error;
     rx_data(16).rate    <= p16_rx_rate;
+    rx_data(16).status  <= p16_rx_status;
     rx_data(16).reset_p <= p16_rx_reset;
     tx_ctrl(16).clk     <= p16_tx_clk;
     tx_ctrl(16).ready   <= p16_tx_ready;
@@ -971,6 +1020,7 @@ gen_p17 : if (PORT_COUNT > 17) generate
     rx_data(17).write   <= p17_rx_write;
     rx_data(17).rxerr   <= p17_rx_error;
     rx_data(17).rate    <= p17_rx_rate;
+    rx_data(17).status  <= p17_rx_status;
     rx_data(17).reset_p <= p17_rx_reset;
     tx_ctrl(17).clk     <= p17_tx_clk;
     tx_ctrl(17).ready   <= p17_tx_ready;
@@ -994,6 +1044,7 @@ gen_p18 : if (PORT_COUNT > 18) generate
     rx_data(18).write   <= p18_rx_write;
     rx_data(18).rxerr   <= p18_rx_error;
     rx_data(18).rate    <= p18_rx_rate;
+    rx_data(18).status  <= p18_rx_status;
     rx_data(18).reset_p <= p18_rx_reset;
     tx_ctrl(18).clk     <= p18_tx_clk;
     tx_ctrl(18).ready   <= p18_tx_ready;
@@ -1017,6 +1068,7 @@ gen_p19 : if (PORT_COUNT > 19) generate
     rx_data(19).write   <= p19_rx_write;
     rx_data(19).rxerr   <= p19_rx_error;
     rx_data(19).rate    <= p19_rx_rate;
+    rx_data(19).status  <= p19_rx_status;
     rx_data(19).reset_p <= p19_rx_reset;
     tx_ctrl(19).clk     <= p19_tx_clk;
     tx_ctrl(19).ready   <= p19_tx_ready;
@@ -1040,6 +1092,7 @@ gen_p20 : if (PORT_COUNT > 20) generate
     rx_data(20).write   <= p20_rx_write;
     rx_data(20).rxerr   <= p20_rx_error;
     rx_data(20).rate    <= p20_rx_rate;
+    rx_data(20).status  <= p20_rx_status;
     rx_data(20).reset_p <= p20_rx_reset;
     tx_ctrl(20).clk     <= p20_tx_clk;
     tx_ctrl(20).ready   <= p20_tx_ready;
@@ -1063,6 +1116,7 @@ gen_p21 : if (PORT_COUNT > 21) generate
     rx_data(21).write   <= p21_rx_write;
     rx_data(21).rxerr   <= p21_rx_error;
     rx_data(21).rate    <= p21_rx_rate;
+    rx_data(21).status  <= p21_rx_status;
     rx_data(21).reset_p <= p21_rx_reset;
     tx_ctrl(21).clk     <= p21_tx_clk;
     tx_ctrl(21).ready   <= p21_tx_ready;
@@ -1086,6 +1140,7 @@ gen_p22 : if (PORT_COUNT > 22) generate
     rx_data(22).write   <= p22_rx_write;
     rx_data(22).rxerr   <= p22_rx_error;
     rx_data(22).rate    <= p22_rx_rate;
+    rx_data(22).status  <= p22_rx_status;
     rx_data(22).reset_p <= p22_rx_reset;
     tx_ctrl(22).clk     <= p22_tx_clk;
     tx_ctrl(22).ready   <= p22_tx_ready;
@@ -1109,6 +1164,7 @@ gen_p23 : if (PORT_COUNT > 23) generate
     rx_data(23).write   <= p23_rx_write;
     rx_data(23).rxerr   <= p23_rx_error;
     rx_data(23).rate    <= p23_rx_rate;
+    rx_data(23).status  <= p23_rx_status;
     rx_data(23).reset_p <= p23_rx_reset;
     tx_ctrl(23).clk     <= p23_tx_clk;
     tx_ctrl(23).ready   <= p23_tx_ready;
@@ -1132,6 +1188,7 @@ gen_p24 : if (PORT_COUNT > 24) generate
     rx_data(24).write   <= p24_rx_write;
     rx_data(24).rxerr   <= p24_rx_error;
     rx_data(24).rate    <= p24_rx_rate;
+    rx_data(24).status  <= p24_rx_status;
     rx_data(24).reset_p <= p24_rx_reset;
     tx_ctrl(24).clk     <= p24_tx_clk;
     tx_ctrl(24).ready   <= p24_tx_ready;
@@ -1155,6 +1212,7 @@ gen_p25 : if (PORT_COUNT > 25) generate
     rx_data(25).write   <= p25_rx_write;
     rx_data(25).rxerr   <= p25_rx_error;
     rx_data(25).rate    <= p25_rx_rate;
+    rx_data(25).status  <= p25_rx_status;
     rx_data(25).reset_p <= p25_rx_reset;
     tx_ctrl(25).clk     <= p25_tx_clk;
     tx_ctrl(25).ready   <= p25_tx_ready;
@@ -1178,6 +1236,7 @@ gen_p26 : if (PORT_COUNT > 26) generate
     rx_data(26).write   <= p26_rx_write;
     rx_data(26).rxerr   <= p26_rx_error;
     rx_data(26).rate    <= p26_rx_rate;
+    rx_data(26).status  <= p26_rx_status;
     rx_data(26).reset_p <= p26_rx_reset;
     tx_ctrl(26).clk     <= p26_tx_clk;
     tx_ctrl(26).ready   <= p26_tx_ready;
@@ -1201,6 +1260,7 @@ gen_p27 : if (PORT_COUNT > 27) generate
     rx_data(27).write   <= p27_rx_write;
     rx_data(27).rxerr   <= p27_rx_error;
     rx_data(27).rate    <= p27_rx_rate;
+    rx_data(27).status  <= p27_rx_status;
     rx_data(27).reset_p <= p27_rx_reset;
     tx_ctrl(27).clk     <= p27_tx_clk;
     tx_ctrl(27).ready   <= p27_tx_ready;
@@ -1224,6 +1284,7 @@ gen_p28 : if (PORT_COUNT > 28) generate
     rx_data(28).write   <= p28_rx_write;
     rx_data(28).rxerr   <= p28_rx_error;
     rx_data(28).rate    <= p28_rx_rate;
+    rx_data(28).status  <= p28_rx_status;
     rx_data(28).reset_p <= p28_rx_reset;
     tx_ctrl(28).clk     <= p28_tx_clk;
     tx_ctrl(28).ready   <= p28_tx_ready;
@@ -1247,6 +1308,7 @@ gen_p29 : if (PORT_COUNT > 29) generate
     rx_data(29).write   <= p29_rx_write;
     rx_data(29).rxerr   <= p29_rx_error;
     rx_data(29).rate    <= p29_rx_rate;
+    rx_data(29).status  <= p29_rx_status;
     rx_data(29).reset_p <= p29_rx_reset;
     tx_ctrl(29).clk     <= p29_tx_clk;
     tx_ctrl(29).ready   <= p29_tx_ready;
@@ -1270,6 +1332,7 @@ gen_p30 : if (PORT_COUNT > 30) generate
     rx_data(30).write   <= p30_rx_write;
     rx_data(30).rxerr   <= p30_rx_error;
     rx_data(30).rate    <= p30_rx_rate;
+    rx_data(30).status  <= p30_rx_status;
     rx_data(30).reset_p <= p30_rx_reset;
     tx_ctrl(30).clk     <= p30_tx_clk;
     tx_ctrl(30).ready   <= p30_tx_ready;
@@ -1293,6 +1356,7 @@ gen_p31 : if (PORT_COUNT > 31) generate
     rx_data(31).write   <= p31_rx_write;
     rx_data(31).rxerr   <= p31_rx_error;
     rx_data(31).rate    <= p31_rx_rate;
+    rx_data(31).status  <= p31_rx_status;
     rx_data(31).reset_p <= p31_rx_reset;
     tx_ctrl(31).clk     <= p31_tx_clk;
     tx_ctrl(31).ready   <= p31_tx_ready;